добавит компонент Container, поправит структуру проекта, добавит логотип
