+++ src/components/Footer/Footer.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import Container from "../Container";
+import s from './Footer.module.css';
+
+export const Footer = () => {
+    return (
+        <footer className={s.root}>
+            <Container>
+                <div className={s.footerWrap}>
+                    Coded with
+                    <span className={s.heart} />
+                    by You
+                </div>
+            </Container>
+        </footer>
+    );
+};
