+++ src/components/Footer/Footer.module.css
@@ -7,15 +7,6 @@
     align-items: center;
 }
 
-.container {
-    position: relative;
-    margin: 0 50px;
-    color: #444;
-    max-width: 1180px;
-    width: 100%;
-    overflow: hidden;
-}
-
 .footerWrap {
     width: 100%;
     display: flex;
@@ -35,24 +26,6 @@
     font-size: 24px;
     color: red;
     animation: spin 2s infinite linear
-}
-
-@media (max-width: 1024px) {
-    .container {
-        margin: 0 25px;
-    }
-}
-
-@media (max-width: 768px) {
-    .container {
-        margin: 0 14px;
-    }
-}
-
-@media (max-width: 560px) {
-    .container {
-        margin: 0 8px;
-    }
 }
 
 @keyframes spin {
