+++ src/components/Footer/index.js
@@ -1,18 +1,3 @@
-import React from 'react';
-import s from './styles.module.css';
-
-const Footer = () => {
-    return (
-        <footer className={s.root}>
-            <div className={s.container}>
-                <div className={s.footerWrap}>
-                    Coded with
-                    <span className={s.heart} />
-                    by You
-                </div>
-            </div>
-        </footer>
-    );
-};
+import {Footer} from "./Footer";
 
 export default Footer;
